Refactor into separate class files
