+++ Cashier.java
@@ -1,0 +1,20 @@
+// Клас, представящ касиер в магазина
+public class Cashier {
+    int id;
+    String name;
+    double salary;
+
+    public Cashier(int id, String name, double salary) {
+        this.id = id;
+        this.name = name;
+        this.salary = salary;
+    }
+
+    public double getSalary() {
+        return salary;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
