+++ ProductCategory.java
@@ -1,0 +1,5 @@
+// Категории на стоки: Хранителни и Нехранителни
+public enum ProductCategory {
+    FOOD,
+    NON_FOOD
+}
